pyproject.toml
[project]
name = "expence-trecker-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "requests>=2.33.1",
    "uvicorn>=0.42.0",
]
